+++ relay-cardinality/src/statsd.rs
@@ -1,4 +1,4 @@
-use relay_statsd::{CounterMetric, TimerMetric};
+use relay_statsd::{CounterMetric, HistogramMetric, TimerMetric};
 
 /// Counter metrics for the Relay Cardinality Limiter.
 pub enum CardinalityLimiterCounters {
@@ -8,7 +8,7 @@ pub enum CardinalityLimiterCounters {
     ///  - `scope`: The scope of check operation.
     #[cfg(feature = "redis")]
     Rejected,
-    /// Incremented for every loaded Redis set.
+    /// Incremented for every redis cardinality check.
     ///
     /// This metric is tagged with:
     ///  - `scope`: The scope of check operation.
@@ -20,14 +20,6 @@ pub enum CardinalityLimiterCounters {
     ///  - `scope`: The scope of check operation.
     #[cfg(feature = "redis")]
     RedisHashCheck,
-    /// Incremented for every hash added to the internal Redis sets.
-    ///
-    /// Note: there are multiple Redis commands executed for every update.
-    ///
-    /// This metric is tagged with:
-    ///  - `scope`: The scope of check operation.
-    #[cfg(feature = "redis")]
-    RedisHashUpdate,
 }
 
 impl CounterMetric for CardinalityLimiterCounters {
@@ -39,8 +31,6 @@ impl CounterMetric for CardinalityLimiterCounters {
             Self::RedisRead => "cardinality.limiter.redis.read",
             #[cfg(feature = "redis")]
             Self::RedisHashCheck => "cardinality.limiter.redis.hash.check",
-            #[cfg(feature = "redis")]
-            Self::RedisHashUpdate => "cardinality.limiter.redis.hash.update",
         }
     }
 }
@@ -54,6 +44,24 @@ impl TimerMetric for CardinalityLimiterTimers {
     fn name(&self) -> &'static str {
         match self {
             CardinalityLimiterTimers::CardinalityLimiter => "cardinality.limiter.duration",
+        }
+    }
+}
+
+pub enum CardinalityLimiterHistograms {
+    /// Redis stored set cardinality.
+    ///
+    /// This metric is tagged with:
+    ///  - `scope`: The scope of check operation.
+    #[cfg(feature = "redis")]
+    RedisSetCardinality,
+}
+
+impl HistogramMetric for CardinalityLimiterHistograms {
+    fn name(&self) -> &'static str {
+        match *self {
+            #[cfg(feature = "redis")]
+            Self::RedisSetCardinality => "cardinality.limiter.redis.set_cardinality",
         }
     }
 }